created useCardStore; flipcard func cannot work
